fix another issue with indentation
